Improve MLIR compatibility with CI builds, add caching of LLVM (#326)
* Switching to HTTPS for Git clone, easier on build agent and works around SSH key update annoyance
* Fix local abbreviation to git command
* Install dependencies for LLVM build i.e. Ninja
* Change kscmlir to use local repository. On integration builds we want to check the one it pulled.
* Pull out LLVM branch commit id to \knossos-ksc\etc\llvm-branch.txt to allow cleaner specification

+++ test/builds/build_and_test_kscmlir.sh
@@ -13,12 +13,18 @@ export LLVM_BUILD="$LLVM_BUILD"
 
 # Clone Knossos and build ksc-mlir
 # This configuration requires clang+lld
-if [ ! -d knossos-ksc ]; then
-  git clone https://github.com/microsoft/knossos-ksc.git
+#if [ ! -d knossos-ksc ]; then
+#  git clone https://github.com/microsoft/knossos-ksc.git
+#fi
+#cd knossos-ksc/mlir
+
+if [ ! -f "SYNTAX.md" ]; then
+  echo "SYNTAX.md isn't present, is this a Knossos repository? Aborting"
+  return 1
 fi
-cd knossos-ksc/mlir
+cd mlir
 KSC_MLIR_SRC="$(pwd)"
-git pull
+#git pull
 rm -rf build
 mkdir build && cd build
 cmake -G Ninja "$KSC_MLIR_SRC" \
